Добавил сценарий оформления брони, но без оплаты. И еще добавил поля для даты. Ну и по мелочи всякое

+++ app/src/main/java/com/pogreb/leasingshift/feature/carinfo/domain/entity/CarInfo.kt
@@ -14,7 +14,7 @@ data class CarInfo(
     val brand: Brand,
     val media:  List<Media>,
     val transmission: Transmission,
-    val price: Double,
+    val price: Long,
     val location: String,
     val color: Color,
     val bodyType: BodyType,
